+++ BoidsBuisnessLogic/Data/Point.cs
@@ -6,8 +6,8 @@ namespace BoidsBuisnessLogic.Data
 {
     public struct Point
     {
-        public double X {  get; set; }
-        public double Y { get; set; }
+        public float X {  get; set; }
+        public float Y { get; set; }
 
         public override string ToString()
         {
