+++ BoidsBuisnessLogic/Samples/SampleDataGenerator.cs
@@ -12,12 +12,12 @@ namespace BoidsBuisnessLogic.Samples
         {
             var result = new List<Point>();
 
-            for (double x = 0; x < 8; x = x + 0.001)
+            for (float x = 0; x < 8; x = x + 0.001f)
             {
                 result.Add(new Point()
                 {
                     X = x,
-                    Y = Math.Cos(x + i),
+                    Y = (float)Math.Cos(x + i),
                 });
             }
